# Improve CSR matrix column fancy indexing