--- a/cupyx/scipy/sparse/compressed.py
+++ b/cupyx/scipy/sparse/compressed.py
@@ -526,7 +526,7 @@ class _compressed_sparse_matrix(sparse_data._data_matrix,
         _, N = self._swap(*self.shape)
         M = len(idx)
         new_shape = self._swap(M, N)
-        if M == 0:
+        if self.nnz == 0 or M == 0:
             return self.__class__(new_shape)
 
         row_nnz = cupy.diff(self.indptr)
@@ -544,10 +544,10 @@ class _compressed_sparse_matrix(sparse_data._data_matrix,
     def _minor_index_fancy(self, idx):
         """Index along the minor axis where idx is an array of ints.
         """
-        M, N = self._swap(*self.shape)
-        k = idx.size
-        new_shape = self._swap(M, k)
-        if k == 0:
+        M, _ = self._swap(*self.shape)
+        N = idx.size
+        new_shape = self._swap(M, N)
+        if self.nnz == 0 or N == 0:
             return self.__class__(new_shape)
 
         if idx.size * M < self.nnz: