… cupyx/scipy/sparse/_index.py