Refactoring and performance improvement.
- Current master
```
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [10935 11560], format = csr, density = 0.01, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU: 2769.116 us   +/-226.974 (min: 2721.915 / max: 6617.697) us     GPU-0: 2825.199 us   +/-225.069 (min: 2779.136 / max: 6652.928) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [10935 11560], format = csr, density = 0.5, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU:68223.292 us   +/-59.720 (min:68180.333 / max:68386.785) us     GPU-0:71651.549 us   +/-46.909 (min:71607.300 / max:71771.133) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [ 8969  1577  7771 ...  9153 10603  7396], format = csr, density = 0.01, dtype = float32, n_rows = 15000,
n_cols = 15000
cupy.sparse         :    CPU: 3428.466 us   +/-93.903 (min: 3396.345 / max: 4953.144) us     GPU-0: 3732.885 us   +/-93.084 (min: 3701.760 / max: 5243.904) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [ 8969  1577  7771 ...  9153 10603  7396], format = csr, density = 0.5, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU:77174.855 us   +/-48.431 (min:77112.003 / max:77281.749) us     GPU-0:94480.476 us   +/-60.847 (min:94386.177 / max:94581.757) us
```

- This PR
```
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [ 1130 14233], format = csr, density = 0.01, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU: 1028.886 us   +/-128.377 (min: 1010.069 / max: 4746.847) us     GPU-0: 1036.010 us   +/-128.627 (min: 1016.832 / max: 4759.552) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [ 1130 14233], format = csr, density = 0.5, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU:29992.919 us   +/-49.895 (min:29916.344 / max:30075.689) us     GPU-0:30001.092 us   +/-50.172 (min:29927.423 / max:30084.097) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [13557 13391  4935 ... 11843  2951  4273], format = csr, density = 0.01, dtype = float32, n_rows = 15000,
n_cols = 15000
cupy.sparse         :    CPU: 1271.756 us   +/-50.393 (min: 1253.454 / max: 2627.985) us     GPU-0: 1291.087 us   +/-50.253 (min: 1272.832 / max: 2642.944) us
- name = minor_column_fancy_indexing, row = slice(None, None, None), col = [13557 13391  4935 ... 11843  2951  4273], format = csr, density = 0.5, dtype = float32, n_rows = 15000, n_cols = 15000
cupy.sparse         :    CPU:37127.139 us   +/-39.527 (min:37064.716 / max:37207.042) us     GPU-0:38817.043 us   +/-36.087 (min:38761.471 / max:38897.663) us
```
cc: @cjnolet 
<details>

- Reproducer

```py
import cupy
from cupy import random
import cupyx.time


def index(array, row, col):
    if min is None:
        return array[row]
    else:
        return array[row, col]


def run(name, row, col, format, density, dtype, n_rows, n_cols):
    a = cupy.sparse.random(n_rows, n_cols, format=format, density=density)
    a = a.astype(dtype)
    print('- name = {}, row = {}, col = {}, format = {}, density = {}, '
          'dtype = {}, n_rows = {}, n_cols = {}'.format(
              name, row, col, format, density, dtype, n_rows, n_cols))
    perf = cupyx.time.repeat(
        index,
        (a, row, col),
        name='cupy.sparse',
        max_duration=1,
    )
    print(perf)


if __name__ == '__main__':
    params = [
        ('minor_column_fancy_indexing', 'csr', slice(None), random.randint(0, 14999, size=2)),
        ('minor_column_fancy_indexing', 'csr', slice(None), random.randint(0, 14999, size=4000)),
    ]
    for name, format, row, col in params:
        for density in [0.01, 0.5]:
            for dtype in ['float32']:
                for n_rows in [15000]:
                    for n_cols in [15000]:
                        run(name, row, col, format, density, dtype, n_rows, n_cols)
```